+++ lib/v2/domain-shared/ui_constants.dart
@@ -1,0 +1,1 @@
+const double defaultCardBorderRadious = 12.0;
